restrutured Settings so that it auto-re-renders
