+++ frontend/src/components/Manager/AdjustPTO.jsx
@@ -3,6 +3,7 @@ import useAuth from "../../hooks/useAuth";
 import axios from "axios";
 import { ToastContainer, toast } from "react-toastify";
 
+
 export default function AdjustPTO({ getAllEmployees, el }) {
   const [user, token] = useAuth();
